fix: restructure components
