+++ src/components/shared/Footer/index.tsx
